requirements.txt
streamlit>=1.36.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyMuPDF>=1.24.0
nltk>=3.8.1
python-dotenv>=1.0.1
requests>=2.32.3
tqdm>=4.66.4
pandas>=2.2.2
numpy>=1.26.4
